Use run with mount instead of copy /output
We currently copy /output into the target image so we can do the
install of the artifacts. Although we then delete it, it's still
in that intermediate layer. That is mildly wasteful. Instead,
use the buildkit extensions to mount /output from the builder
image. These are present by default in modern Docker.

This shaves 147M off of the built image size of zuul/zuul,
bringing it from 868M to 721M in a local build.

Change-Id: Ie0b2b92bbf712a259b3e69291268bf08c37f89df