Checkout playbook branch for roles running on tag
For tag triggered builds the executor was falling back on the project
default branch when checking out roles, as the 'branch' build parameter
was None in those cases.

Instead we will now use the playbook's branch as a fallback for
resolving the branch for a role when no branch is provided.

Change-Id: I4569de1d011b301d9d28059e603149c6becebd10