Add runtime graphs to web
Add a job runtimes page to the web app.  This uses the new
build-times endpoint to supply its data.

A link to the runtimes page is added to the build page, so that
when investigating a build result, the graph of the most recent
runtimes for similar builds can easily be seen.

Future modifications can add the ability to adjust the time period
of the graph.

Change-Id: I7549fef6bce89ab2f18e452052155c922bde8be1