Switch quickstart to docker compose v2
We want to use podman to run our containers, but the podman-compose
tool has been problematic to use in OpenDev and has also caused
some confusion for users.  It also seems the podman project is
favoring declaring systems using k8s syntax rather than compose,
which is putting the podman-compose tool future in question.

This change updates the quickstart to use "docker compose" (ie,
docker-compose v2) instead of podman-compose.

Because we want to continue using speculative container testing
which does not work for quay.io images when using docker, we will
use the podman backend with docker compose.  This is what we
are moving toward in OpenDev.  We hope that it is the best of both
worlds.

Users likely don't need such a complex setup, so we will adjust
the documentation to use "docker compose" but not mention anything
about podman.  That will remain a test implementation detail,
at least as long as we think they are sufficiently compatible that
we can get away with testing something slightly different than
our instructions.

This switches to packaged git-review for simplicity because
pip install does not work in noble.

Fedora and OpenSuse installation instructions are similarly updated.

Red Hat / CentOS instructions are removed because this author does
not have a way to verify package names.  Their return by someone
knowledgable with that is welcome.

Change-Id: I238f6db178af86fa7203203993c90bd8d1371712
Depends-On: https://review.opendev.org/923083
Depends-On: https://review.opendev.org/925916
Co-Authored-By: Clark Boylan <clark.boylan@gmail.com>