Add metrics for the number of online components
We already report the number of executors and mergers online, but
since all components are scalable now, report the number of schedulers,
web, fingergw, (and launchers).

The web and scheduler metric hierarchies already existed and were
singular, but executors and mergers were plural.  We should be
consistent, and in the future, move from one to the other, but since
we are not currently, metrics for scheduler and web are put in the
existing singular hierarchy (to avoid confusion from having "scheduler"
and "schedulers").  New hierarchies are created as plural.

The launcher metric is added, but, in accordance with our current
practice, it will remain undocumented until the nodepool-in-zuul work
is complete.

Change-Id: I15fa0316c637971e2d7da195b960a62248f540c2