Fix test_supercedent_github_circular_deps_closed race (2)
This test is subject to race conditions where if the initial
pipeline status is not created before the nologs check, the nologs
assertion will trip on it.  Make sure the pipeline objects are
created before the main part of the test.

This replaces an earlier fix with a slightly more sophisticated
fix.

Change-Id: I7023eac9a6c0942c2c46f393adc04988e3e420bb