web: Use a select filter for pipelines and queues on status page
On the PipelineOverview and PipelineDetails page, we know the set of
valid pipeline and queue names. With this change we suggest those to the
user when filtering in the form of the type-ahead select filter we
already use in other places. This should make it clearer to users that
they can filter for multiple values by selecting more options from the
filter suggestions.

To retain the "wildcard filtering" feature, this change removes the
separate "fuzzy-search" filter type and instead adds a new flag for
this, so we can enable this for other filter types, too.

Change-Id: I395d2b51441c02ffcf8a07770e42791cbaf62f0e