Ignore file matches on changes to include-vars and playbooks
If a user changes the configuration of a job, we automatically run
that job, regardless of file matchers, so that the job is self-testing
without needing to specify the zuul.yaml file explicitly in a file
matcher.

The same should be true for include-vars as well as playbooks.  They
both should be considered part of the job configuration, and since
we have their paths, we can check them against the change file list
and behave in the same way as the rest of the job configuration.

Change-Id: I3f834230558bf842b5b55f7b4fc55f75f87c9d23