Reuse configuration objects
This change switches out the unparsed config cache in favor of a
new config object cache which holds parsed config objects.

These objects are used in all tenants, and when a tenant is
reconfigured, only the objects from the changed files are replaced
in the cache.

When a layout is created, it is assembled from the relevant objects
in the cache (filtering for the various settings on the tenant
project config).

The rough overview of configloader caching is now:

* File cache (serialized json in ZK, populated from merger cat jobs)
* Config object cache (python objects in memory)

None of which have any tenant-specific information.

This does make one intentional behavioral change: the load-classes
for trusted and untrusted projects are currently identical except
that untrusted projects are not configured to load pipelines.  The
sequence of events when creating a static and dynamic layout is
slightly different, and the load-classes end up filtering pipelines
out from untrusted projects before we check for them and add the
PipelineNotPermitted configuratione error.  But in the dynamic path,
we check for that error first.

Because both code paths are much closer in behavior with this change,
we need to resolve this discrepancy.  In this change, it is resolved
in favor of adding the configuration error to the static code path.
That means that the load-classes for trusted and untrusted projects
are the same, and we rely solely on the check in addProjectBranchConfig
to either add the pipeline or add the error.  This method is called
in both the static and dynamic paths.

Change-Id: I6a63848f8478ed087ae493981fd1dc81dfff842d