Improve job result colors on status page
Some users reported that builds resulting in a POST_FAILURE are hard to
spot on the new status page as they are using the same orange color like
CANCELED builds. To better distinguish those, don't mark CANCELED builds
as failure and use a grey color instead.

To go one step further, we decided to align the result colors as
follows:

- Red: Errors in the job that are actionable by the user, like:
  failure, timed_out, merge_conflict
- Orange: Errors which may be system errors or actionable by the user,
  like: post_failure, node_failure, retry_limit
- blue/grey: Termination without actionable error, like:
  skipped, canceled

Additionally, fix the QUEUED result which points to a non-existing CSS
color and add the missing MERGE_CONFLICT result to the color mapping.

Change-Id: Ic26d6aebb27323332fa683a3440e5d7a67fa081d