Remove tenant references from project parsing
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

This change removes tenant references from the parsing of project
and project-template stanzas.

Change-Id: Ibbc70e30d607e0ac52e1c90214ee2925341732fc