Stop copying ProjectConfig objects when adding to layout
The previous change added a condition where we would copy a
ProjectConfig object while adding it to the layout if it specified
a project name that was not fully resolved.  This was a workaround
for cases where a project stanza with a short project name might
reference different projects in different tenants.

Instead of doing that, let's remove the requirement that a
ProjectConfig object know the actual project it's configuring.
Rather like jobs, ProjectConfigs will be applied to whatever projects
they should apply to in that tenant.

To accomplish this, we will continue to store project config objects
on the layout in a dictionary keyed by the project canonical names
that the objects apply to.  But when we add objects to that dict,
we will let the caller tell us what the project cname any given
project definition is for.  So a single ProjectConfig object may refer
to gerrit/foo in one tenant, and github/foo in another tenant.
When we decide what ProjectConfigs are relevant for a change,
look them up by the project cname of the change and pass that cname
to the ProjectConfig branch matcher so that we have full context
of how the ProjectConfig is being used.

This also resolves the only reason we are creating duplicate
ProjectConfig objects for regex-based project stanzas, so those are
eliminated as well.

Change-Id: I2e034c299761fb3ae70b33b2104e09cea2269bdc