Check for nodeset label perms in layout
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

This change moves the check for whether a nodeset label is allowed
by a tenant to the point where we add the nodeset object to the
layout.

Since anonymous nodesets can be embedded in jobs, and anonymous
jobs can be embedded in project-templates and projects, we also
need to check those.

Change-Id: I74ff4f96ff84f16326c72f6db7d0b3d4b7b60893