Check for job timeouts in layout
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
job's context in a tenant must happen when the layout is
constructed.

This change moves the check for whether a job timeout is allowed
by a tenant to the point where we add the job to the layout.

Since multiple job-related checks are needed for all three ways
of adding a job to the layout, they are refactored into a new
method.

Change-Id: Ifdc93a42dd19d313ffe1364da62082ea36f0b344