Resolve required-projects in layout
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
job's context in a tenant must happen when the layout is
constructed.

This change moves the resolution of project references in
required-projects to later stages.  The value is checked when the
job is added to the layout, but not saved.  The final values are
saved during job freezing (with setBase and applyVariant).

Change-Id: I30fced71f40c6f7afad887a6850a47b73744296f