Resolve include-vars projects in layout
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
job's context in a tenant must happen when the layout is
constructed.

This change moves the resolution of project references in
include-vars to the layout.  Since we aren't freezing and thawing
anymore, it also drops the use of tuples in the Job class.

The fromDict method was not used and is removed.

Change-Id: I339bfb4b6cad7fee28378c83d85f8295fd4e299c