Check allowed-projects in layout
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
job's context in a tenant must happen when the layout is
constructed.

This change moves the check for whether an allowed-project is
in the tenant to the point where we add the job to the layout.

Change-Id: I72dc6a18af99289fc9b603b2c3c504835c6dd454