Move ignore_allowed_projects to job freeze
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
job's context in a tenant must happen when the layout is
constructed.

This moves the ignore_allowed_projects check from the job
construction to job freezing.  At this time we have the info
needed to set the flag: whether the job is attached to a PPC in
a trusted project.

Change-Id: I991238bb09e27b05375b771895cd52ac6019e314