Move post_review setting to job freezing
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

This moves the setting of the post_review flag to job freezing
where we have information about whether secrets are being set
from trusted or untrusted projects.

To speed the check whether a project is trusted or not in the
tenant, a new cached isTrusted method is added.

Change-Id: I1f5c1349abd8cb23bf8e5d7c8894b5bbe03b11b2