Implement launcher relative priority handling
This change adds support for revision the relative priority of nodeset
requests similar to how this is already possible for Nodepool.

Similar to build/merge requests we are not allowed to modify the nodeset
request after it has been created. To still allow revision of the
relative priority we will create a 'revision' subnode, which contains
the updated relative priority of the request.

The treecache was updated to handle the 'revision' subnode and update
the nodeset request accordingly.

Change-Id: I8b154cd3626b1e0251c142e37b0a2e2ed7e9c0d5