Move base job check to layout
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

This moves the check for whether it is permitted to define a
base job to the point where the job is added to the layout.

Change-Id: Ied8f8b923a4499b14f9e375b98371756943c7f5f