Resolve role references when freezing
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

When we construct a job with role references, those references
previously were made directly to canonical projects in the tenant.
This change moves the resolution of those references to later
during the freezing process.

Change-Id: I838cfb0ad997d5be1b2df144ff01124180780664