Set playbook trusted flag when freezing
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

Previously, the trusted flag for a playbook was set based on whether
the source context was trusted.  This change sets it when the playbook
is frozen as part of the act of job freezing.  That decouples it from
the source context used when parsing config.

Change-Id: I48812e64230384a4a044820ed17419a73647db9d