Remove trusted flag from source context
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

Since whether a project is trusted or not is associated with
the tenant context, it is not appropriate for the source context
to have a trusted flag.  This change removes that flag to ensure
that we always use the flag from the tenant project config instead.

Change-Id: I7f243c84aea4cd84cae51567e15605f2948d39e7