Determine job branch matchers when freezing
As part of an effort to fully re-use parsed config objects, we
need to remove any references to tenant configuration when
parsing the objects.  All behaviors that are affected by the
project's context in a tenant must happen when the layout is
constructed.

This moves the branch matcher selection logic for a job from
job parsing to job freezing.  This way we store information
about the parse context, then choose whether to use explicit
or implied branch matchers when we have a tenant context during
freezing.

Change-Id: I0c6baa6cc7b5d4a0a130bc8f09cdbbccf61b3669