Move some pipeline attributes to manager
To reimagine the pipeline as an abstract config object, move some
attributes such as the relative priority queues, change list, and
pipeline state to be attributes of the pipeline manager.

Since the test infrastructure now has a check for unexpected
exceptions, remove the pipeline exception counter which was only
used for that purpose.

Change-Id: Ia8c861faf29a29694c7c0e31c593e042955d6f7d