Remove the manager attribute from pipelines
This moves the model.Pipeline class closer to being a tenant and
layout-independent configuration object by removing its "manager"
attribute.  Instead, we will consider the Pipeline class an attribute
of the manager.  The PipelineManager will be the root class for all
information about a pipeline.  It will hold the PipelineState and
PipelineSummary ZKObjects, as well as the Pipeline object which
describes its configuration.

This does change a lot of references, probably more than strictly
necessary.  But since, over the years, we built up several helper
methods to get to a pipeline's queues from various objects, it was
becoming unclear how to actually get to the objects we are interested
in after this change.  So to avoid increasing the length of our
object reference chains, this change rips off the bandage and
directly changes many references.

Change-Id: Ied771c56e4fc5a1e032737811a3818168ef36fce