Allow creating (sharded) ZKObjects w/o makepath
So far ZKObject and its sharded equivalent also created the parent path
in case it did not exist. Since we are sometimes working with cached
data, the information we have could be stale.

If we now want to create ZKObjects as subnodes of existing nodes we
could cause problems when e.g. the parent node was deleted in the
meantime and we re-create as part of the sub-ZKObject without any
content.

To prevent this data race, we make this behavior configurable per
ZKObject by introducing a `makepath` attribute which is set to True by
default.

Change-Id: Ib17e40cd5d664bfc625d83c742d76dabd2dd7a8c