AWS: Add imdsv2 support
This adds a feature deferred from the nodepool import.

In nodepool we named this setting "imdsv2" but here it is renamed
to "imds-http-tokens" to reduce the delta from the aws api.  This
option sets the "HttpTokens" key of the "MetadataOptions" instance
key.

A bug with the provider setting inheritance was fixed where we would
potentially use a None value rather than an actually set value.

Change-Id: Ia57b121e904caae7a854b35a449669d23b4f09b3