Fix pause event verification test race
The SQL and MQTT drivers both verify that pause events are resumed
properly by checking that the resume event timestamp is greater than the
pause event timestamp. These timestamps appear to have one second of
resolution (confirmed for the database, not 100% certain of mqtt) which
means that on a really fast node you can run all of the dependent jobs
in less than a second and resume the pause parent job without ever
ticking the clock over to the next second. When that happens we get
errors like this:

  Traceback (most recent call last):
    File "/home/zuul/src/opendev.org/zuul/zuul/tests/unit/test_connection.py", line 287, in test_sql_results
      check_results()
    File "/home/zuul/src/opendev.org/zuul/zuul/tests/unit/test_connection.py", line 224, in check_results
      self.assertGreater(
    File "/usr/lib/python3.11/unittest/case.py", line 1271, in assertGreater
      self.fail(self._formatMessage(msg, standardMsg))
    File "/usr/lib/python3.11/unittest/case.py", line 703, in fail
      raise self.failureException(msg)
  AssertionError: datetime.datetime(2025, 3, 11, 0, 5, 50) not greater than datetime.datetime(2025, 3, 11, 0, 5, 50)

We fix that by having the dependent jobs pause for 1 second to ensure
that the parent is always resumed at least one second later. This has
the downside of causing all unittests using these jobs to slow by one
second. An alternative would be to change the timestamp comparison to a
greater than or equal check, but this seems like strict in what it is
asserting. We want to ensure that the children have run allowing things
to unpause.

Change-Id: Ic681e7f970826b925a768f2644cfc972cb1c4106