Fix pause event verification test race
The SQL and MQTT drivers both verify that pause events are resumed
properly by checking that the resume event timestamp is greater than the
pause event timestamp. These timestamps appear to have one second of
resolution (confirmed for the database, not 100% certain of mqtt) which
means that on a really fast node you can run all of the dependent jobs
in less than a second and resume the pause parent job without ever
ticking the clock over to the next second. When that happens we get
errors like this:

  Traceback (most recent call last):
    File "/home/zuul/src/opendev.org/zuul/zuul/tests/unit/test_connection.py", line 287, in test_sql_results
      check_results()
    File "/home/zuul/src/opendev.org/zuul/zuul/tests/unit/test_connection.py", line 224, in check_results
      self.assertGreater(
    File "/usr/lib/python3.11/unittest/case.py", line 1271, in assertGreater
      self.fail(self._formatMessage(msg, standardMsg))
    File "/usr/lib/python3.11/unittest/case.py", line 703, in fail
      raise self.failureException(msg)
  AssertionError: datetime.datetime(2025, 3, 11, 0, 5, 50) not greater than datetime.datetime(2025, 3, 11, 0, 5, 50)

To fix this we more carefully release the pausing job, wait for it to
pause, then sleep for one second before releasing the other jobs. This
ensures that the child jobs (start and) complete at least one second
after the pause event. This ensures the resume event time stamp is at
least one second later than the pause event time stamp.

Change-Id: Ic681e7f970826b925a768f2644cfc972cb1c4106