WIP: Implement pre-timeout for pre-run playbooks
This implements a very old TODO for Zuulv3 from pabelanger. We add a
pre-run playbook timeout. This is useful if you want to trigger job
retries due to pre run setup not completing within a reasonable amount
of time.

For backward compatibiltiy and general sanity we continue to count
pre-run execution time against the normal timeout value. This basically
allows you to limit the span of time that pre-run is allowed to consume
from the larger timeout. This is different to the post-timeout behavior.

Change-Id: I33ea67e40553fd9dce3ce5025ebc21e247ce83e2