Only consider live items on dequeue event
So far dequeue events did not distinguish between live and non-live
items. Combined with the fact that we only dequeue a single item per
dequeue event we could end up removing a non-live item and leaving the
actual live item in the queue.

Fix this by only considering live items when handling a dequeue event.

Change-Id: Iffa489735b6e6f08ca10a2060e72f78251880b80