Launcher: implement resource-limits
This implements provider resource limits.  These could be used like
max-servers in nodepool, or, because sections can be defined
centrally with unique providers in different tenants, this may
be able to act somewhat like nodepool's tenant-resource-limits at
the same time.  At least it would allow users to specify tenant
resource limits for a single provider (but not across multiple
providers in a tenant).

Change-Id: Icc2803c28d75f05741965cd6fb57cc0de4ee1299