Avoid reconfig for quick create/delete of branches
So far we've always considered a branch protected when got a branch
related event, but could no longer figure out if the referenced branch
was created or not.

This can lead to unnecessary reconfigurations in case a unprotected
branch is created/deleted before any of the associated events are
processed.

It looks like this case can be optimized by ignoring branch-creation
events for which the branch no longer exists as it's safe to assume that
we never loaded any config from that branch in the first place (even if
it was protected).

Change-Id: Ibf16c7bcb17fa0979f305bc442fd352bbf24cde4