Allow Moto exception when listing instances
Our test cases check logs for unexpected exceptions and the test case
fails if exceptions are found. Unfortunately, moto doesn't appear to be
super thread safe when listing instances. Jinja2 ends up raising for a
changing dictionary size when Moto puts together an instance listing
response for us. Since this isn't all that different to a network
problem problem or cloud error zuul-launcher seems to handle it just
fine but our test case fails anwyay.

Add this exception type to okay_tracebacks on the test case that seems
to trip it to allow things to pass in this case. Maybe we want to see if
moto can fix it upstream using a copy of dictionaries they iterate over?

Change-Id: If41122819be864ca8d072a265183f9bb7987e4ba