Retry moto describe_instance exceptions
Our test cases check logs for unexpected exceptions and the test case
fails if exceptions are found. Unfortunately, moto doesn't appear to be
super thread safe when listing instances. Jinja2 ends up raising for a
changing dictionary size when Moto puts together an instance listing
response for us and reservations have changed in the process. This
causes test cases to log a RuntimeError then fail on the unexpected
exception.

Address this by patching describe_instances to call itself up to 10
times in succession returning the first successful result and catching
the RuntimeError otherwise.

Change-Id: If41122819be864ca8d072a265183f9bb7987e4ba