Add Windows live log streaming
This adds support for live log streaming on Windows hosts.

The daemon is written in C#, but is structured exactly like the
Python daemon for Posix, so that any changes to one can be easily
translated to the other.

Like the Python/Posix version, it is necessary to vendor some
Powershell and C# code from Ansible and the ansible.windows
collection.  Minimal changes were made to these (in fact, much
more minimal than the changes to the command.py module for the
Posix version).  All changes involve the word "Zuul" for easy
identification.  Most of the work to write out the console
log files is in a new file, Ansible.Zuul.Win.Common.cs which
appears only in Zuul.

Becasue it's possible to run the Posix log streaming daemon
in WSL, and it's also possible to run both "win_command" and
"command" tasks on the same host, outside and inside of WSL,
the Windows log streaming daemon (win_zuul_console) runs on
a different port, so that both may be running on the same host
at the same time.  This is more straightforward than trying
to have both daemons be aware of the file paths of the other.

I did make an attempt to create a test environment with wine,
dotnet, and powershell all running under linux that we could use
in the gate, however something about how Ansible links the C#
modules is different enough to render that nonfunctional.
Perhaps someday it will work; until then, we will have to use our
best efforts to maintain this and manually test with a Windows
VM.  The procedure for running the remote tests is the same for
Windows as it is for Posix.

A copy of the "command" remote streaming test is added for windows,
but disabled in the gate due to test resources.

Change-Id: I35c220c1774b1943165075c3730236c78b95b18d