Support mariadb statement timeouts
MariaDB has a different method of handling statement timeouts compared
to MySQL.  To support both, we need sqlalchemy to know whether it is
connected to a mysql or mariadb server, which means we need our users
to specify the mariadb dialect in their dburi.

Add documentation indicating that they should do so, and add support
for the mariadb dialect and the specific method of handling statement
timeouts.

Switch the quickstart dburi to use the mariadb dialect since that's
what we recommend now.

Note that mariadb specifies its timeouts in seconds, unlike other
dialects which use milliseconds.

Change-Id: Idb9df7d34c988190153ca5899a90d44e38b2fcca