Use mysql query hint regardless of project
We use a mysql/mariadb index hint to speed up some queries where
the query planner examines too many rows.  We avoid that hint when
there are enough columns involved in the select clause that allow
it to narrow the results with indexes.

The project column was one that we thought would narrow results
sufficiently, but it appears that may not be the case.  Including
the project in the standard web ui's queries for builds or buildsets
without the hint causes the queries to take longer that the query
timeout in opendev when searching for nova.  Meanwhile, adding the
hint results in fast results, even for smaller projects.

Therefore, add the query hint regardless of whether the project is
specified.  We continue to omit the hint if project is specified
for the build-times query, since that behaves differently and does
not appear to be exhibiting problems.

Change-Id: I7a9d74ffe67bec991e9726298c3b1c6f08ec4e5b