Run a deduplicated job if it matched any change
So far when deduplicating jobs we only checked if the jobs can be
deduplicated from a config perspective. However, we also need to
consider if any of the jobs matched a change.

To fix this we'll make sure to run the deduplicated job if any of the
jobs matches the change.

Change-Id: I8a0354e999e3499d85360d3bf28737ef3ae3e9b5