Fix promote event handling for github changes
Currently, the following exception is raised whenever Zuul tries to
handle a promote event for a github change:

    2025-03-19 15:12:26,572 ERROR zuul.Scheduler: Exception in management event:
    Traceback (most recent call last):
      File "/opt/zuul/lib/python3.11/site-packages/zuul/scheduler.py", line 2754, in _process_management_event
        self._doPromoteEvent(event)
      File "/opt/zuul/lib/python3.11/site-packages/zuul/scheduler.py", line 1975, in _doPromoteEvent
        raise Exception("Unable to find %s,%s" % (number, patchset))
    Exception: Unable to find 223,8eb32a030e97eb4bef5e33543e4fd801a77954ab

This is caused by a type mismatch when looking up the change from the
event data. While the event data contains a string containing the
<number>,<patchset> combination, the `item_change.number` for the
enqueued item is usually an integer. To fix this, convert the number
and patchset from the enqueued change objects to string to properly
compare them with the event data.

A more profound solution would be to use the getChangeKey() method from
the project source, which is also used to handle enqueue and dequeue
events. This method does a similar type conversion underneath to look up
change object. However, to use the getChangeKey() method requires the
project name to be available which is currently not included in the
promote event data. Thus, as a long-term improvement it might make sense
to adapt the promote event handling to be closer to the enqueue/dequeue
event handling.

Change-Id: I62cfce82fc750a59b9ccd183fa2702447a45e9ef