fix(merger): fix file comment line mapping
File comment mapping gives incorrect file comment line mapping
when the change contains large deletion (greater than 9 lines)
next to the comment line as the diff_re would only match a
single digit for the span. And the matching would continue
for commits under the specified commit in the git log if diff
matching is not found for the specified commit, which is also
incorrect logic since we should only be looking for diff
matching for the specified commit.

Hence the fix to this bug would be:
1. make span of the original text in the diff_re match for
multiple digits
2. stop diff matching after scanning the diff lines for the
specified commit

Change-Id: I722f4fde3f7908bc0f4a5e13c096d05af76ec645