Try to use mariadb in unittest again
The last time we tried this we had to revert due to flaky testing that
resulted. Now that we're aware mariadb vs mysql dialects may be
impactful we try again explicitly using the mariadb dialect in testing.

We switch zuul-nox and children over to debian-bookworm instead of
ubuntu-jammy so that we get python 3.11 and mariadb 10.11 on the host.
ubuntu-jammy has mariadb 10.6 which may be the source of the per test
db schema problems.

I think before we land this we will want to do much rechecking in order
to determine if any flaky tests are due to database instability.
Previously the db issues seemed to be around test isolation iirc.

Change-Id: I2f85a07cae2e7ba6566806ec70dbc3ac36a77f10