Try to use mariadb in unittest again
The last time we tried this we had to revert due to flaky testing that
resulted. Now that we're aware mariadb vs mysql dialects may be
impactful we try again explicitly using the mariadb dialect in testing.

I think before we land this we will want to do much rechecking in order
to determine if any flaky tests are due to database instability.
Previously the db issues seemed to be around test isolation iirc.

Change-Id: I2f85a07cae2e7ba6566806ec70dbc3ac36a77f10