Make MySQL test fixture more robust
While debugging MariaDB connectivity failures there were three things
that I noticed we could do to the MySQL test fixture to make it more
robust that should be applicable whether or not we switch to MariaDB.

1. Try to connect to the database within our try except block so that we
   can catch and raise exceptions caused by the initial connection and
   not just the queries.
2. Log the host, port, and database name info (previously it was just
   the database name that was logged).
3. Don't use flush privileges unnecessarily. It is suspected that this
   may contribute to MariaDB connection problems. But even if it doesn't
   these flushes shouldn't be required and would only cause the db to
   do extra work potentially slowing down the test environment.

In this process I also noticed that pymysql library has a debug flag
that allows us to capture very verbose debug info. Add support for
toggling that flag within the fixture but leave it off by default (due
to its verbosity).

Change-Id: I3993a84a605752e6a247be2c76a6da507879c21a