Move zuul-nox-py311-multi-scheduler to the experimental pipeline
This job seems to always timeout at the 2 hour point. This means you
have to wait at least 2 hours for any results to report to gerrit if
this job has run. Considering it has been flaky and never really worked
anyway lets move it to experiemental where it can be picked up again
later if there is renewed interest.

Change-Id: Ieef6d4df60d71984222e02b0adf94f3933b31c3d