Add logging of kubectl process group id to executor
When debugging which kubectl processes below to which jobs it is useful
to know the process group id for those processes in addition to the port
that is being bound to. Add that info to the log entry that we report
with the port number.

I also rewrite the log.info() call to pass the interpolation arguments
as arguments to the logging call rather than using str.format(). This
ensures that we only interpolate if we're going to record the log
message in the first place.

Change-Id: I40ea98e6eb938266fe9a94e630b011cd450be3e1