web: Pass auth header to log endpoint if serving from the same origin
If we serve log files from the same origin as the API, pass the auth
header. This is to make sure we don't have to do authentication for the
log file requests again if the same reverse proxy is handling the
requests and authorization. This must be enabled explictly by setting
AUTH_LOG_REQUESTS during js build time.

Change-Id: I32139671db194c1407146911f677ad2516159e11