web: Pass auth header to log endpoint if serving from the same origin
If we serve log files from the same origin as the API, pass an
authorization header with the log file request if available.

This is to make sure we don't have to do authentication for the
log file requests again if the same reverse proxy is handling the
requests and authorization.

This must be enabled explictly by setting a new zuul-web config flag
`web.auth_log_file_requests` to true.

Change-Id: I32139671db194c1407146911f677ad2516159e11