Add error list to error accumulator
Currently the configloader error accumulator operates on a single
ConfigErrors object -- every error that is encountered is added
to that object which is attached to the resulting tenant.  But
in order to cache complete objects without reparsing them (and
regenerating errors), we need to store the errors with the objects.

To facilitate that (in a later change), we switch the local error
accumulator to operate on a simple list, and allow the extended
local accumulators to override that list.  That way we can change
where the errors are stored as we proceed through the parsing
process.

Eventually, some errors will be stored on the tenant we are building,
and others will be stored in the config object cache along with
their objects, and added into the tenant later.

For now, we still store them all on a single list and just add that
list to the tenant when done.

Change-Id: I91fe3c6f8c0ed7a12205db25763f75918d59c9e4