Store object errors separately
This stores object errors on the ParsedConfig object we create,
then adds those errors to the layout when we add the resulting
config objects.  This will allow us to make the parseConfig method
tenant independent, and allow us to store the object parse errors
along with the objects when we start caching ParsedConfigs.

Change-Id: I3da5a364e90d428f55b093ea60dd2dcc818d0349