Refactor unprotected branch tests
These tests verify behavior by inspecting the unparsed branch cache
which we want to remove.  Update the tests to verify behavior by
examining the actual resulting layout rather than the cache that was
used to create that layout.  They should work now and in the future
despite changes to configloader caching.

Change-Id: Ia469865efb1f2d4d9f8f8ef6dc39015ad53e853f