Fix race in queue tests
These tests can race the node request cleanup method, which could
leave a pipeline state object with the read-only flag set on the
pipeline.  This is normally cleared in the scheduler by refreshing
the state.  Do that before getting the queue so that this this
a better simulation of what the scheduler will actually do, and
corrects the discrepancy introduced by the race.

Change-Id: I2516bc5734b28f24c83aa36f7e3cd0b535863a77