WIP Make Gerrit Connection Replication Aware
If Zuul receives replication events and is configured to wait for them
(by setting a positive timeout value) the gerrit connection will now
wait for replication to report completion before processing the related
events. This should allow Zuul deployments to rely on their replicas for
Zuul git operations.

This is still very much a work in progress. The three new test cases
pass but I've written a number of TODOs to make this better. Also we
need release notes and documentation.

Change-Id: Ie1dfccac1f5a088547caa999352811ad6842f3e5