Make Gerrit Connection Replication Aware
If Zuul receives replication events and is configured to wait for them
(by setting a positive timeout value) the gerrit connection will now
wait for replication to report completion before processing the related
events. This should allow Zuul deployments to rely on their replicas for
Zuul git operations.

While Zuul's test suite seems to indicate this works there are some
things to keep in mind. We can only associate replication and primary
events using project,ref tuples. This means that if many updates are
made to the same project,ref simultaneously Zuul must wait for all of
them to complete before processing any of the primary events.
Additionally, Gerrit replicas may have different performance and Zuul
may not talk to all of them. A potential followup improvement would be
to specify a matcher to match against replica names to limit the
replicas to only those that impact Zuul.

Change-Id: Ie1dfccac1f5a088547caa999352811ad6842f3e5